Correct regular expressions
Several of them used strings instead of JS regular expressions. This is allowed because the string is passed to the RegExp constructor, but the syntax is subtly different. This caused one regexp to be invalid and some others to not match correctly.

This converts them to normal JS regex syntax, which seems to work correctly for all of them. It also corrects the one that was using the wrong check on the result, and removes the /g modifier from the single match-or-no-match checks where it's not needed.